app(scraper): refactor
